Extract bitpix/tform into their own modules
